# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface RestaurantDTO {
  restaurant_id: number;
  name: string;
}

export interface MenuItemDTO {
  id: number;
  name: string;
  price: number;
  description?: string | null;
  available: boolean;
  restaurant?: RestaurantDTO;
  subgroup?: string | null;
}

export interface CreateMenuItemDTO {
  name: string;
  price: number;
  description?: string | null;
  restaurant_id: number;
  available: boolean;
  subgroup?: string | null;
}

export interface OrderDTO {
  id: number;
  date?: string;
  extra_fields?: string | null;
  menu_items: MenuItemDTO[];
  restaurant: RestaurantDTO;
  table_number: number;
  status: string;
}

export interface CreateOrderDTO {
  restaurant_id: number;
  menu_items_ids: number[];
  extra_fields?: string | null;
  table_number: number;
}

export interface OrderStatusDTO {
  status: string;
}

export interface BillDTO {
  restaurant_id: number;
  table_number: number;
}