# lib/api-zod/src/generated/types/billDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface BillDTO {
  restaurant_id: bigint;
  table_number: bigint;
}