# lib/api-zod/src/generated/types/createMenuItemDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemDTO {
  name: string;
  price: number;
  description?: string | null;
  restaurant_id: bigint;
  available: boolean;
  subgroup?: string | null;
}