# lib/api-zod/src/generated/types/createOrderDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderDTO {
  restaurant_id: bigint;
  menu_items_ids: bigint[];
  extra_fields?: string | null;
  table_number: bigint;
}