# lib/api-zod/src/generated/types/menuItemDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { RestaurantDTO } from "./restaurantDTO";

export interface MenuItemDTO {
  id: bigint;
  name: string;
  price: number;
  description?: string | null;
  available: boolean;
  restaurant?: RestaurantDTO;
  subgroup?: string | null;
}