# lib/api-zod/src/generated/types/orderDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemDTO } from "./menuItemDTO";
import type { RestaurantDTO } from "./restaurantDTO";

export interface OrderDTO {
  id: bigint;
  date?: Date;
  extra_fields?: string | null;
  menu_items: MenuItemDTO[];
  restaurant: RestaurantDTO;
  table_number: bigint;
  status: string;
}