# lib/api-zod/src/generated/types/restaurantDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scaneat Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface RestaurantDTO {
  restaurant_id: bigint;
  name: string;
}